implement shane's comments
