+++ selfdrive/ui/layouts/main.py
@@ -2,14 +2,15 @@ import pyray as rl
 from enum import IntEnum
 import cereal.messaging as messaging
 from openpilot.system.ui.lib.application import gui_app
+from openpilot.system.ui.widgets import Widget
 from openpilot.selfdrive.ui.layouts.sidebar import Sidebar, SIDEBAR_WIDTH
 from openpilot.selfdrive.ui.layouts.home import HomeLayout
 from openpilot.selfdrive.ui.layouts.settings.settings import SettingsLayout, PanelType
 from openpilot.selfdrive.ui.onroad.augmented_road_view import AugmentedRoadView
 from openpilot.selfdrive.ui.ui_state import device, ui_state
-from openpilot.system.ui.widgets import Widget
 from openpilot.selfdrive.ui.layouts.onboarding import OnboardingWindow
-from openpilot.selfdrive.ui.body.layouts import BodyLayout, BodySidebar, BODY_SIDEBAR_HEIGHT
+from openpilot.selfdrive.ui.body.layouts.onroad import BodyLayout
+from openpilot.selfdrive.ui.body.widgets.sidebar import BodySidebar, BODY_SIDEBAR_HEIGHT
 
 
 class MainState(IntEnum):
@@ -30,11 +31,9 @@ class MainLayout(Widget):
     self._prev_onroad = False
 
     # Initialize layouts
+    self._layouts = {MainState.HOME: HomeLayout(), MainState.SETTINGS: SettingsLayout(), MainState.ONROAD: AugmentedRoadView()}
     if self._is_body:
-      self._layouts = {MainState.HOME: BodyLayout(), MainState.SETTINGS: SettingsLayout(), MainState.ONROAD: BodyLayout()}
-      self._sidebar.set_visible(False)
-    else:
-      self._layouts = {MainState.HOME: HomeLayout(), MainState.SETTINGS: SettingsLayout(), MainState.ONROAD: AugmentedRoadView()}
+      self._layouts[MainState.HOME] = BodyLayout()
 
     self._sidebar_rect = rl.Rectangle(0, 0, 0, 0)
     self._content_rect = rl.Rectangle(0, 0, 0, 0)
@@ -57,14 +56,14 @@ class MainLayout(Widget):
     self._sidebar.set_callbacks(on_settings=self._on_settings_clicked,
                                 on_flag=self._on_bookmark_clicked,
                                 open_settings=lambda: self.open_settings(PanelType.TOGGLES))
-    self._layouts[MainState.HOME]._setup_widget.set_open_settings_callback(lambda: self.open_settings(PanelType.FIREHOSE))
+    if not self._is_body:
+      self._layouts[MainState.HOME]._setup_widget.set_open_settings_callback(lambda: self.open_settings(PanelType.FIREHOSE))
     self._layouts[MainState.HOME].set_settings_callback(lambda: self.open_settings(PanelType.TOGGLES))
     self._layouts[MainState.SETTINGS].set_callbacks(on_close=self._set_mode_for_state)
-    if self._is_body:
-      self._layouts[MainState.HOME].set_click_callback(self._on_onroad_clicked)
-      self._layouts[MainState.ONROAD].set_click_callback(self._on_onroad_clicked)
+    self._layouts[MainState.HOME].set_click_callback(self._on_onroad_clicked)
 
-    device.add_interactive_timeout_callback(self._set_mode_for_state)
+    if not self._is_body:
+      device.add_interactive_timeout_callback(self._set_mode_for_state)
 
   def _update_layout_rects(self):
     if self._is_body:
@@ -77,7 +76,7 @@ class MainLayout(Widget):
       self._content_rect = rl.Rectangle(self._rect.x + x_offset, self._rect.y, self._rect.width - x_offset, self._rect.height)
 
   def _handle_onroad_transition(self):
-    if ui_state.started != self._prev_onroad:
+    if ui_state.started != self._prev_onroad and not self._is_body:
       self._prev_onroad = ui_state.started
       self._set_mode_for_state()
 
@@ -121,10 +120,9 @@ class MainLayout(Widget):
                                  self._rect.width, self._rect.height - BODY_SIDEBAR_HEIGHT) if self._sidebar.is_visible else None
       self._layouts[self._current_mode].set_parent_rect(parent_rect)
       self._layouts[self._current_mode].render(self._rect)
-      if self._sidebar.is_visible:
-        self._sidebar.render(self._sidebar_rect)
     else:
-      if self._sidebar.is_visible:
-        self._sidebar.render(self._sidebar_rect)
       content_rect = self._content_rect if self._sidebar.is_visible else self._rect
       self._layouts[self._current_mode].render(content_rect)
+
+    if self._sidebar.is_visible:
+        self._sidebar.render(self._sidebar_rect)
