+++ selfdrive/ui/mici/layouts/home.py
@@ -186,11 +186,10 @@ class MiciHomeLayout(Widget):
       self._date_label.set_position(version_pos.x + self._version_label.text_width + 10, version_pos.y)
       self._date_label.render()
 
-      if ui_state.CP is not None and not ui_state.CP.notCar:
-        self._branch_label.set_max_width(gui_app.width - self._version_label.text_width - self._date_label.text_width - 32)
-        self._branch_label.set_text(" " + ("release" if release_branch else self._version_text[1]))
-        self._branch_label.set_position(version_pos.x + self._version_label.text_width + self._date_label.text_width + 20, version_pos.y)
-        self._branch_label.render()
+      self._branch_label.set_max_width(gui_app.width - self._version_label.text_width - self._date_label.text_width - 32)
+      self._branch_label.set_text(" " + ("release" if release_branch else self._version_text[1]))
+      self._branch_label.set_position(version_pos.x + self._version_label.text_width + self._date_label.text_width + 20, version_pos.y)
+      self._branch_label.render()
 
       if not release_branch:
         # 2nd line
